Make test dependencies optional
Dependencies for tests are only installed when the `test` feature is
requested. Also made version to be dynamic, based on `__version__` in
`__init__.py`. Also, the case discrepancy between the project name and
the module directory name was causing problems (on Linux), so I added a
module name section for flit to ensure it finds the module.